+++ bibdesk/BDSKTableView.m
@@ -322,7 +322,7 @@ static char BDSKTableViewFontDefaultsObservationContext;
     if ([self canDelete]) {
         NSUInteger originalNumberOfRows = [self numberOfRows];
         // -selectedRow is last row of multiple selection, no good for trying to select the row before the selection.
-        NSUInteger selectedRow = [[self selectedRowIndexes] firstIndex];
+        NSInteger selectedRow = [[self selectedRowIndexes] firstIndex];
         [[self dataSource] tableView:self deleteRowsWithIndexes:[self selectedRowIndexes]];
         [self reloadData];
         NSUInteger newNumberOfRows = [self numberOfRows];
@@ -340,7 +340,7 @@ static char BDSKTableViewFontDefaultsObservationContext;
                 }
             } else {
                 // Don't try to go past the new # of rows
-                selectedRow = MIN(selectedRow - 1, newNumberOfRows - 1);
+                selectedRow = MIN(selectedRow - 1, (NSInteger)newNumberOfRows - 1);
                 
                 // Skip all unselectable rows if the delegate responds to -tableView:shouldSelectRow:
                 if ([[self delegate] respondsToSelector:@selector(tableView:shouldSelectRow:)]) {
